Move copyOneImage, imageCopier, and exclusive callees, to copy/single.go
Only moves unchanged code (in the same order), should not change behavior.
