+++ copy/single_test.go
